Remove old custom view API
