+++ src/app/toolbar/ViewSelect.svelte
@@ -1,6 +1,5 @@
 <script lang="ts">
-  import { customViews, customViewsV2 } from "src/lib/stores/custom-views";
-  import { Builder } from "src/builder";
+  import { customViews } from "src/lib/stores/custom-views";
   import type { ViewDefinition } from "src/types";
 
   import ViewItemList from "./ViewItemList.svelte";
@@ -26,21 +25,7 @@
       case "gallery":
         return "layout-grid";
       default:
-        const createView = $customViewsV2[type];
-
-        if (createView) {
-          const view = createView();
-          return view.getIcon();
-        }
-
-        const builder = $customViews[type];
-
-        if (builder) {
-          const view = new Builder();
-          builder(view);
-          return view.icon ?? "";
-        }
-        return "";
+        return $customViews[type]?.()?.getIcon() ?? "";
     }
   }
 </script>
