+++ src/lib/stores/custom-views.ts
@@ -1,8 +1,4 @@
-import type { ProjectView, ProjectViewV2 } from "../../builder";
+import type { ProjectViewV2 } from "../../custom-view-api";
 import { writable } from "svelte/store";
 
-export const customViews = writable<
-  Record<string, (view: ProjectView) => void>
->({});
-
-export const customViewsV2 = writable<Record<string, () => ProjectViewV2>>({});
+export const customViews = writable<Record<string, () => ProjectViewV2>>({});
